Fix a warning about missing Tls 1.1/1.2 protocols
A warning could be issued, even if the protocols are present and a
correct .Net Framework is installed.

It's enough to trigger the warning just when SSL is not supported - and
now it's absent from the .NET Core, so the warning is always there in
PowerShell Core.

It's because the "try {}" block tests the whole following line:
"(...) 3072 -bor 768 -bor 192 -bor 48" and fails on "-bor 48" part.

Why care about SSL 3.0, when we have Tls 1.1/1.2 encryption set?

What's changed is that now foreach loop tries to set a highest
encryption available by testing the protocols one by one, from SSL until
Tls 1.2. "catch {}" block is triggered when a specific protocol is
absent and the final "if" block checks the final selected encryption
(if it's Tls 1.2 or 1.1).

The solution works for PowerShell 2.x/.Net 2.x.