Set SecurityProtocol to all available protocols
Simplified version of a script to fix a warning about missing Tls 1.x
protocols, even if they're present in the system. The warning was
triggered on PowerShell 6.1.x (.Net Core 2.x) because of a missing SSL
protocol. It should be ignored, as newer protocols Tls 1.x are
available.

The commit is a simplified version of the previous fix. It's passing the
integer value of all available protocols to
[System.Net.ServicePointManager]:SecurityProtocol, instead of just
selected one.